LPD-72635 Move all the export logic to a ReverseIterator to parse the Document Library references
